test: isolate cyclic foreign key data branch regression
